# docs/openapi/openapi.json
{
  "components" : {
    "schemas" : {
      "AccountPasswordRequest" : {
        "properties" : {
          "passwordHash" : {
            "minLength" : 1,
            "type" : "string",
            "writeOnly" : true
          }
        },
        "required" : [ "passwordHash" ],
        "type" : "object"
      },
...
      "AccountSummary" : {
        "properties" : {
          "email" : {
            "format" : "email",
            "type" : "string"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "locale" : {
            "enum" : [ "ca", "es", "en" ],
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "platformRoles" : {
            "items" : {
              "enum" : [ "AGILITYHUB_ADMIN" ],
              "type" : "string"
            },
            "type" : "array",
            "uniqueItems" : true
          }
        },
        "required" : [ "email", "id", "locale", "name", "platformRoles" ],
        "type" : "object"
      },
      "ApiError" : {
        "properties" : {
          "code" : {
            "type" : "string"
          },
          "details" : {
            "additionalProperties" : true,
            "type" : "object"
          },
          "message" : {
            "type" : "string"
          },
          "traceId" : {
            "type" : "string"
          }
        },
        "required" : [ "code", "details", "message", "traceId" ],
        "type" : "object"
      },
...
      "ImpersonationRequest" : {
        "properties" : {
          "reason" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "ImpersonationTokenResponse" : {
        "properties" : {
          "expiresAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "token" : {
            "type" : "string"
          }
        },
        "required" : [ "expiresAt", "token" ],
        "type" : "object"
      },
...
      "PasswordRequest" : {
        "properties" : {
          "current" : {
            "format" : "password",
            "type" : "string",
            "writeOnly" : true
          },
          "new" : {
            "format" : "password",
            "minLength" : 1,
            "type" : "string",
            "writeOnly" : true
          },
          "repeat" : {
            "format" : "password",
            "minLength" : 1,
            "type" : "string",
            "writeOnly" : true
          }
        },
        "required" : [ "new", "repeat" ],
        "type" : "object"
      },
      "PlatformAccountRequest" : {
        "properties" : {
          "email" : {
            "format" : "email",
            "minLength" : 1,
            "type" : "string"
          },
          "locale" : {
            "enum" : [ "ca", "es", "en" ],
            "minLength" : 1,
            "type" : "string"
          },
          "name" : {
            "minLength" : 1,
            "type" : "string"
          },
          "passwordHash" : {
            "description" : "Existing Learn bcrypt hash, preserved on import",
            "type" : "string",
            "writeOnly" : true
          }
        },
        "required" : [ "email", "locale", "name" ],
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "bearer" : {
        "bearerFormat" : "JWT",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
  "info" : {
    "title" : "AgilityHub Core API",
    "version" : "v1"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/v1/accounts/{id}/password" : {
      "put" : {
        "description" : "R-01-12. Authenticated learn client with accounts:write scope, or AGILITYHUB_ADMIN (S17). Idempotent; never returns the hash.",
        "operationId" : "accountPassword",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/AccountPasswordRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Password synchronized"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Bad Request"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Forbidden"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Not Found"
          },
          "409" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Conflict"
          },
          "422" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Unprocessable Entity"
          },
          "429" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Too Many Requests",
            "headers" : {
              "Retry-After" : {
                "description" : "Seconds before retrying",
                "schema" : {
                  "format" : "int32",
                  "type" : "integer"
                }
              }
            }
          },
          "4XX" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Client error"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Internal Server Error"
          },
          "501" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Not Implemented"
          },
          "5XX" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Server error"
          }
        },
        "summary" : "Synchronize an existing Learn password hash",
        "tags" : [ "identity" ]
      }
    },
...
    "/api/v1/me/password" : {
      "put" : {
        "description" : "Account token, never impersonated. R-01-05: current is required only when a password already exists.",
        "operationId" : "password",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PasswordRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Password changed"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "PASSWORD_TOO_SHORT, PASSWORD_MISMATCH, PASSWORD_COMPROMISED"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "INVALID_CREDENTIALS for current password"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "FORBIDDEN for impersonated accounts"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Not Found"
          },
          "409" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Conflict"
          },
          "422" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Unprocessable Entity"
          },
          "429" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Too Many Requests",
            "headers" : {
              "Retry-After" : {
                "description" : "Seconds before retrying",
                "schema" : {
                  "format" : "int32",
                  "type" : "integer"
                }
              }
            }
          },
          "4XX" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Client error"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Internal Server Error"
          },
          "501" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Not Implemented"
          },
          "5XX" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Server error"
          }
        },
        "summary" : "Set or change the current account password",
        "tags" : [ "identity" ]
      }
    },
...
    "/api/v1/members/{id}/impersonation-token" : {
      "post" : {
        "description" : "ADMIN of the same club, never an impersonated token. R-01-09. Target lookup and grant issuance are E1-T04.",
        "operationId" : "impersonation",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ImpersonationRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ImpersonationTokenResponse"
                }
              }
            },
            "description" : "Impersonation JWT and UTC expiry"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Bad Request"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "IMPERSONATION_DENIED"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "NOT_FOUND within the current club"
          },
          "409" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Conflict"
          },
          "422" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Unprocessable Entity"
          },
          "429" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Too Many Requests",
            "headers" : {
              "Retry-After" : {
                "description" : "Seconds before retrying",
                "schema" : {
                  "format" : "int32",
                  "type" : "integer"
                }
              }
            }
          },
          "4XX" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Client error"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Internal Server Error"
          },
          "501" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Not Implemented"
          },
          "5XX" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Server error"
          }
        },
        "summary" : "Create a non-refreshable member impersonation token",
        "tags" : [ "identity" ]
      }
    },
    "/api/v1/platform/accounts" : {
      "post" : {
        "description" : "R-01-12. Authenticated learn client with accounts:write scope, or AGILITYHUB_ADMIN (S17). Idempotent.",
        "operationId" : "createAccount",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PlatformAccountRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AccountSummary"
                }
              }
            },
            "description" : "Public account including its ID for the Learn mapping"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Bad Request"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Forbidden"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Not Found"
          },
          "409" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "EMAIL_EXISTS"
          },
          "422" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Unprocessable Entity"
          },
          "429" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Too Many Requests",
            "headers" : {
              "Retry-After" : {
                "description" : "Seconds before retrying",
                "schema" : {
                  "format" : "int32",
                  "type" : "integer"
                }
              }
            }
          },
          "4XX" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Client error"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Internal Server Error"
          },
          "501" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Not Implemented"
          },
          "5XX" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Server error"
          }
        },
        "summary" : "Provision a global account",
        "tags" : [ "identity" ]
      }
    },
...
  },
  "security" : [ {
    "bearer" : [ ]
  } ],
  "servers" : [ {
    "url" : "https://core.agilitydoghub.com"
  }, {
    "url" : "http://localhost:8080"
  } ]
}